% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AASeq.R
\name{AASeq}
\alias{AASeq}
\title{Finding amino acid sequence from codons}
\usage{
AASeq(codons)
}
\arguments{
\item{codons:}{An object expected too be a vector}
}
\value{
amine acid sequence: a string containing the corresponding amino acid sequence
}
\description{
Finding amino acid sequence from codons
}
\examples{
AASeq2() -> FSYCFSYCLS__LS_WLPHRLPHRLPQRLPQRITNSITNSITKRMTKRVADGVADGVAEGVAEG
}
