% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_histogram.R
\name{amino_histogram}
\alias{amino_histogram}
\title{Generate histogram of amino acids}
\usage{
amino_histogram(amino_sequence)
}
\arguments{
\item{amino_sequence}{An amino acid sequence string}
}
\value{
histogram_ggplot
}
\description{
Generate histogram of amino acids
}
