% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dna_sequence.R
\name{create_dna_sequence}
\alias{create_dna_sequence}
\title{Create DNA sequence}
\usage{
create_dna_sequence(length)
}
\arguments{
\item{length}{Desired length of the DNA sequence}
}
\value{
A string containing a randomly generated DNA sequence
}
\description{
Create DNA sequence
}
\examples{
create_dna_sequence(10)
}
