% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Turning a DNA sequence into RNA}
\usage{
dna_to_rna(dna_sequence)
}
\arguments{
<<<<<<< HEAD
\item{dna_Sequence:}{A string containing a polynucleotide sequence of DNA.}
=======
\item{dna_sequence:}{A string containing a polynucleotide sequence of DNA.}
>>>>>>> 0048833e96811b948af430df5b708c6792c807e1
}
\value{
dna_sequence: A string containing a polynucleotide sequence of RNA.
}
\description{
Turning a DNA sequence into RNA
}
\examples{
dna_sequence <- "ATCGTT"
dna_to_rna(dna_sequence)
}
