% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_codons.R
\name{find_codons}
\alias{find_codons}
\title{Find codons}
\usage{
find_codons(coding_sequence, start = 1)
}
\arguments{
\item{coding_sequence}{A DNA or RNA sequence as a string.}

\item{start}{Start position of the sequence.}
}
\value{
A list of 3 character strings, where each string is a codon.
}
\description{
Find codons
}
\examples{
dna_sequence <- "ATCGTTACCGTCCGT"
find_codons(dna_sequence)
}
