% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_codon}
\alias{standard_codon}
\title{Standard Codon}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
standard_codon
}
\description{
Vector containing the standard codon and the corresponding aminoacid.
}
\keyword{datasets}
