% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Turning a DNA sequence into RNA}
\usage{
DNA_to_RNA(DNA_sequence)
}
\arguments{
\item{DNA_Sequence:}{A string containing a polynucleotide sequence of DNA.}
}
\value{
RNA_sequence: A string containing a polynucleotide sequence of RNA.
}
\description{
Turning a DNA sequence into RNA
}
\examples{
Input: "ATTG",   Output: "AUUG".
}
